fix(matomo): start cleaning up
